JSDoc updates, type checking on comparison, login updates, user session fix, etc.
Added some small, extra updates to the JSDocs, and also removed an unnecessary tryCatchHelper for a specific method. Only one error can be thrown after all others were accounted for, so I went ahead and put it within a try catch loop for simplicity. Also fixed a bug with the auth variable and checking user session.